test(nexus): refactored test module
